Radek's recent improvements in firing of changes from filesystem broke this test

+++ openide.loaders/test/unit/src/org/openide/loaders/Deadlock59522Test.java
@@ -113,7 +113,13 @@ public class Deadlock59522Test extends NbTestCase implements FileChangeListener 
         }
     }
     
+    private boolean wasHereOnce;
     private void lockMdr() {
+        if (wasHereOnce) {
+            return;
+        }
+        
+        wasHereOnce = true;
         // no more callbacks
         fs.removeFileChangeListener(this);
